==> notebook/04. DatasetCreation.ipynb <==
import pandas as pd
import numpy as np
import itertools

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

## Defining the Target Variables
- Define the time-splits for your train, validation and test data.
- Define the latitude & longitude ranges of your target country, zone, etc. The example is for Turkey.
- Define the floating point precision for yout longitude-latitude values.
- Define after how many fire calls you will consider to be a wild-fire.
- Define after what confidence level (in percentage) you will treat a report as a wild-fire occurence.

TRAIN_UNTIL = 2018
VAL_BETWEEN = (2018, 2020)
TEST_ON = 2020

LAT_RANGE = (36, 42)
LON_RANGE = (26, 45)

PRECISION = 1
MIN_FIRE_RECORDS = 2

CONFIDENCE_THRESHOLD = 80

# LGBM Regressor Model

We zoomed in Turkey and used aggressive aggregation for a simple baseline prediction model.
* Temporal resolution: Monthly
* Spatial resolution: 'PRECISION' Decimal degree ~ ( 10^(2-'PRECISION') km grid)
* Binary Target: At least 'MIN_FIRE_RECORDS' fire readings

fires = pd.read_csv('../data/processed_data/fire_daily.csv.gz', parse_dates=['acq_date'])
fires.head()

- We are pruning the data based on our desired confidence level.
- We are only taking the zones that falls within our target latitude & longitude ranges.

fires = fires[fires.confidence > CONFIDENCE_THRESHOLD]

fires = fires[
        (fires.latitude > LAT_RANGE[0]) & (fires.latitude < LAT_RANGE[1]) & \
        (fires.longitude > LON_RANGE[0]) & (fires.longitude < LON_RANGE[1])]

fires.shape

We are extracting year and month. Then rounding the coordinate pairs with our defined floating point precision.

fires['year'] = fires.acq_date.dt.year
fires['month'] = fires.acq_date.dt.month
fires.latitude = fires.latitude.round(PRECISION)
fires.longitude = fires.longitude.round(PRECISION)

Grouping all the reports based on year, month and coordinates.

fires = fires.groupby(['latitude', 'longitude', 'year', 'month']).size().reset_index()
fires.columns = ['latitude', 'longitude', 'year', 'month', 'fire_cnt']

fires.shape
fires.head()
fires.nunique()

Creating all the possible coordinate pairs with our desired precision. If we only take the coordinates found in the fire dataset, we would train a model that tends to produce false positives. So we should give all possible coordinates and aim for the model to predict the minority class.

lats = np.arange(LAT_RANGE[0], LAT_RANGE[1], 1/(10^PRECISION)).round(PRECISION)
lons = np.arange(LON_RANGE[0], LON_RANGE[1], 1/(10^PRECISION)).round(PRECISION)
years = fires.year.unique()
months = fires.month.unique()

unq_combs = list(itertools.product(lats, lons))
coords = pd.DataFrame(unq_combs, columns=["latitude", "longitude"])

unq_combs = list(itertools.product(years, months))
times = pd.DataFrame(unq_combs, columns=["year", "month"])

coords['one'] = 1
times['one'] = 1

base = pd.merge(coords, times, how='outer', on='one').drop_duplicates()
history = base.merge(fires, how='left', on= ['latitude', 'longitude', 'year', 'month']).drop_duplicates()

Let's see how many reports can be obtained for a coordinate throughout the area.

history = history.fillna(0)
history.fire_cnt.value_counts().head()

We mark areas that have been reported more than our threshold value as wildfire. This is our ground-truth value.

history['fire'] = 1 * (history['fire_cnt'] >= MIN_FIRE_RECORDS)

- For each year, we add the count of wild-fires and the count of fire reports for the last year as new features.
- For each month, we add the count of wild-fires and the count of fire reports for the same month in last year as new features.

yearly = history.groupby(
    ['latitude', 'longitude', 'year'])[['fire_cnt', 'fire']].sum().reset_index()
monthly = history.groupby(
    ['latitude', 'longitude', 'year', 'month'])[['fire_cnt', 'fire']].sum().reset_index()

last_year = yearly.copy()
last_year.year += 1
last_year.columns = ['latitude', 'longitude', 'year', 'fire_cnt_last_year', 'fire_last_year']
last_year.head()

last_year_month = monthly.copy()
last_year_month.year += 1
last_year_month.columns = ['latitude', 'longitude', 'year', 'month', 'fire_cnt_last_year_same_month', 'fire_last_year_same_month']
last_year_month.head()

By shifting the years by one and comparing the two year columns, we calculate the number of fire reports and wild-fires in the previous years for each year.

past = yearly.copy()
past['one'] = 1
past = history[['latitude', 'longitude', 'year', 'one']].drop_duplicates().merge(
    past, on=['latitude', 'longitude', 'one'])
past = past[past.year_x < past.year_y]
past = past.groupby(['latitude', 'longitude', 'year_y'])[['fire_cnt', 'fire']].sum().reset_index()
past.columns = ['latitude', 'longitude', 'year', 'fire_cnt_before', 'fire_before']
past.head(3)

We combine the historical report and wildfire statistics we extracted with the main dataframe.

X = history.merge(past, how='left', on=['latitude', 'longitude', 'year'])
X = X.merge(last_year, how='left', on=['latitude', 'longitude', 'year'])
X = X.merge(last_year_month, how='left', on=['latitude', 'longitude', 'year', 'month'])
X = X.drop(columns='one')

X.head()
X.shape

# Temperature Data
We are reading the temperature data we created from the Berkeley dataset.

temp_df = pd.read_csv('../data/processed_data/temperatures.csv')
temp_df = temp_df.rename(columns={"latitude":"latitude_rounded",
                                 "longitude":"longitude_rounded"})
temp_df["month"] = temp_df["month"].astype(int)
temp_df["year"] = temp_df["year"].astype(int)

Latitude and longitude pairs in the data is formatted like:
- 39.5
- 40.5
- 41.5
So when we merging the temperature data with our wildfire statistics, we should treat the wildfire data with this conversion (only for merging, we will preserve the actual latitude & longitude pairs).

def half_round(number):
    return np.floor(number) + 0.5

X["latitude_rounded"] = half_round(X["latitude"])
X["longitude_rounded"] = half_round(X["longitude"])

temp_df.head()

We are merging the historical wildfire statistics data with the temperature data.

X = pd.merge(X,
             temp_df,
             how="left",
             on=["latitude_rounded", "longitude_rounded", "month", "year"]
            )

We are mean-imputing the temperature values on yearly basis.

# X["temperature_min"] = X.groupby(['year'])["temperature_min"].transform(lambda x: x.fillna(x.mean()))
# X["temperature_max"] = X.groupby(['year'])["temperature_max"].transform(lambda x: x.fillna(x.mean()))
# X["temperature_avg"] = X.groupby(['year'])["temperature_avg"].transform(lambda x: x.fillna(x.mean()))

X.fire_cnt_before.max()

# Train-Test-Val Splitting

train = X[X.year < TRAIN_UNTIL].dropna()
valid = X[(X.year >= VAL_BETWEEN[0]) & (X.year < VAL_BETWEEN[1])]
test = X[X.year == TEST_ON]

train.to_csv('../data/train_val_test_data/train.csv', index=False)
valid.to_csv('../data/train_val_test_data/valid.csv', index=False)
test.to_csv('../data/train_val_test_data/test.csv', index=False)
X.to_csv('../data/train_val_test_data/full_data.csv', index=False)


train.shape, valid.shape, test.shape
train
